# [REG-2191] Speedup game action analysis