@abeizer 2024-11-22
> I was honestly hoping to get this down to single digit seconds for ALL games

This improvement is huge though! Very impressed by the time save